requirements.txt
asgiref==3.2.10
attrs==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
cx-Oracle==8.0.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==1.11.2
django-filter==1.1.0
djangorestframework==3.9.2
EasyProcess==0.3
html5lib==1.1
idna==2.10
inflect==4.1.0
iniconfig==1.1.1
lxml==4.6.1
lxmlbind==2.0
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
packaging==20.4
paramiko==2.7.2
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.1.1
pytest-assume==2.3.3
pytest-metadata==1.10.0
pytest-reportlog==0.1.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlacodegen==2.3.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.11
webencodings==0.5.1
